Ya pintamos con alpha

+++ Engine/src/main/java/es/ucm/gdv/engine/Graphics.java
@@ -3,7 +3,7 @@ package es.ucm.gdv.engine;
 public interface Graphics {
     Image newImage(String name);
     void clear(int color);
-    void drawImage(Image img, Rect scr, Rect clip, int alpha);
+    void drawImage(Image img, Rect scr, Rect clip, float alpha);
     int getWidth();
     int getHeight();
 }
